requirements.txt
pytz
python-dotenv
redis
supabase
curl_cffi
telebot